requirements.txt
fastapi
uvicorn
pytesseract
Pillow
PyMuPDF
python-multipart
